requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
httpx==0.28.1
python-multipart==0.0.20
aiosqlite==0.20.0
pymysql==1.1.1
cryptography==44.0.0
feedparser==6.0.11
APScheduler==3.10.4
openai==1.58.1
